refactor: Move `CompliantSeries`
